究极体

+++ reading-cloud-homepage/src/main/java/cn/zealon/readingcloud/homepage/domain/SearchBookResult.java
@@ -13,5 +13,5 @@ public class SearchBookResult {
     /** 图书列表 */
     private List<SearchBookItem> bookList;
     /** 返回总数 */
-    private Long total;
+    private int total;
 }
